CIT-6 Updating Ferme

+++ src/main/java/com/citronix/service/Implemenation/FermeServiceImp.java
@@ -3,6 +3,7 @@ package com.citronix.service.Implemenation;
 import com.citronix.dto.ferme.FermeRequestDto;
 import com.citronix.dto.ferme.FermeResponseDto;
 import com.citronix.dto.ferme.FermeSearchCriteria;
+import com.citronix.exceptions.FermeNotFoundException;
 import com.citronix.mapper.ferme.FermeMapper;
 import com.citronix.model.Ferme;
 import com.citronix.repository.FermeRepository;
@@ -10,14 +11,13 @@ import com.citronix.repository.Custom.FermeSpecification;
 import com.citronix.service.Interface.IFermeService;
 import lombok.RequiredArgsConstructor;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
 @Service
-@RequiredArgsConstructor(onConstructor = @__(@Autowired))
+@RequiredArgsConstructor
 public class FermeServiceImp implements IFermeService {
     private final FermeMapper fermeMapper;
     private final FermeRepository fermeRepository;
@@ -26,13 +26,14 @@ public class FermeServiceImp implements IFermeService {
     @Override
     public FermeResponseDto create(FermeRequestDto fermeRequestDto) {
         Ferme ferme = fermeMapper.toEntity(fermeRequestDto);
-        fermeRepository.save(ferme);
-        return fermeMapper.toResponseDto(ferme);
+        Ferme savedFerme = fermeRepository.save(ferme);
+        return fermeMapper.toResponseDto(savedFerme);
     }
 
     @Override
     public FermeResponseDto update(Long id, FermeRequestDto fermeRequestDto) {
-        Ferme ferme = fermeRepository.findById(id).orElseThrow(() -> new RuntimeException("Ferme non trouvée"));
+        Ferme ferme = fermeRepository.findById(id)
+                .orElseThrow(() -> new FermeNotFoundException("Ferme non trouvée"));
 
         ferme.setNom(fermeRequestDto.nom());
         ferme.setLocalisation(fermeRequestDto.localisation());
@@ -40,12 +41,14 @@ public class FermeServiceImp implements IFermeService {
         ferme.setDateCreation(fermeRequestDto.dateCreation());
 
         Ferme updatedFerme = fermeRepository.save(ferme);
-
         return fermeMapper.toResponseDto(updatedFerme);
     }
 
     @Override
     public void delete(Long id) {
+        if (!fermeRepository.existsById(id)) {
+            throw new FermeNotFoundException("Ferme non trouvée");
+        }
         fermeRepository.deleteById(id);
     }
 
@@ -64,5 +67,4 @@ public class FermeServiceImp implements IFermeService {
         fermeSpecification.setCriteria(criteria);
         return fermeRepository.findAll(fermeSpecification);
     }
-
 }
